Organize classes into packages
Divide all classes into multiple packages due to the increase number of classes.
