--- src/main/java/BotType.java
@@ -1,4 +1,0 @@
-public enum BotType {
-    JJBA,
-    DIO,
-}
