--- src/main/java/CommandType.java
@@ -1,9 +1,0 @@
-public enum CommandType {
-    ADD,
-    DELETE,
-    LIST,
-    MARK,
-    UNMARK,
-    BOT,
-    EXIT,
-}
