--- src/main/java/ExitCommand.java
@@ -1,9 +1,0 @@
-public class ExitCommand extends Command {
-    public static final String COMMAND_WORD = "bye";
-    public static final String COMMAND_FORMAT = COMMAND_WORD;
-
-    @Override
-    public CommandFeedback execute(TaskList taskList) {
-        return new CommandFeedback(CommandType.EXIT);
-    }
-}
