--- src/main/java/IndexCommand.java
@@ -1,8 +1,0 @@
-public abstract class IndexCommand extends Command {
-    protected static final String COMMAND_ARG = " <valid task no>";
-    protected final int index;
-
-    protected IndexCommand(int index) {
-        this.index = index;
-    }
-}
