--- src/main/java/InvalidArgumentException.java
@@ -1,15 +1,0 @@
-public class InvalidArgumentException extends DukeException {
-    public InvalidArgumentException() {
-        super("", false, true);
-    }
-
-    public InvalidArgumentException(String errorMsg) {
-        super(errorMsg, false, true);
-    }
-
-    public InvalidArgumentException(String errorMsg, boolean isHidden) {
-        super(errorMsg, isHidden, true);
-    }
-
-
-}
