+++ src/main/java/duke/bot/BotType.java
@@ -1,0 +1,6 @@
+package duke.bot;
+
+public enum BotType {
+    JJBA,
+    DIO,
+}
