+++ src/main/java/duke/command/CommandFeedback.java
@@ -1,0 +1,30 @@
+package duke.command;
+
+import duke.task.Task;
+import duke.task.TaskList;
+
+public class CommandFeedback {
+    public final CommandType cType;
+    public final TaskList taskList;
+    public final Task task;
+
+    public CommandFeedback(CommandType cType) {
+        this(cType, null, null);
+    }
+
+    public CommandFeedback(CommandType cType, TaskList taskList) {
+        this(cType, taskList, null);
+    }
+
+    public CommandFeedback(CommandType cType, Task task) {
+        this(cType, null, task);
+    }
+
+    public CommandFeedback(CommandType cType, TaskList taskList, Task task) {
+        this.cType = cType;
+        this.taskList = taskList;
+        this.task = task;
+    }
+
+}
+
