+++ src/main/java/duke/command/CommandType.java
@@ -1,0 +1,11 @@
+package duke.command;
+
+public enum CommandType {
+    ADD,
+    DELETE,
+    LIST,
+    MARK,
+    UNMARK,
+    BOT,
+    EXIT,
+}
