+++ src/main/java/duke/command/DeadlineCommand.java
@@ -1,0 +1,12 @@
+package duke.command;
+
+import duke.task.Deadline;
+
+public class DeadlineCommand extends AddCommand {
+    public static final String COMMAND_WORD = "deadline";
+    public static final String COMMAND_FORMAT = COMMAND_WORD + " <task name> /by dd-MM-yyyy HHmm";
+
+    public DeadlineCommand(Deadline newDeadline) {
+        super(newDeadline);
+    }
+}
