+++ src/main/java/duke/command/IndexCommand.java
@@ -1,0 +1,10 @@
+package duke.command;
+
+public abstract class IndexCommand extends Command {
+    protected static final String COMMAND_ARG = " <valid task no>";
+    protected final int index;
+
+    protected IndexCommand(int index) {
+        this.index = index;
+    }
+}
