+++ src/main/java/duke/command/TodoCommand.java
@@ -1,0 +1,13 @@
+package duke.command;
+
+import duke.task.Todo;
+
+public class TodoCommand extends AddCommand {
+    public static final String COMMAND_WORD = "todo";
+    public static final String COMMAND_FORMAT = COMMAND_WORD + " <task name>";
+
+    public TodoCommand(Todo newTodo) {
+        super(newTodo);
+    }
+}
+
