+++ src/main/java/duke/exception/InvalidArgumentException.java
@@ -1,0 +1,17 @@
+package duke.exception;
+
+public class InvalidArgumentException extends DukeException {
+    public InvalidArgumentException() {
+        super("", false, true);
+    }
+
+    public InvalidArgumentException(String errorMsg) {
+        super(errorMsg, false, true);
+    }
+
+    public InvalidArgumentException(String errorMsg, boolean isHidden) {
+        super(errorMsg, isHidden, true);
+    }
+
+
+}
