+++ src/main/java/duke/task/Todo.java
@@ -1,0 +1,21 @@
+package duke.task;
+
+public class Todo extends Task {
+    public Todo(String name) {
+        super(name);
+    }
+
+    public Todo(String name, boolean isMark) {
+        super(name, isMark);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+
+    @Override
+    public String toData() {
+        return "T|" + super.toData();
+    }
+}
